services/pipeline/requirements.txt
anthropic>=0.39.0
openai>=1.54.0
google-generativeai>=0.8.0
google-genai
python-dotenv>=1.0.0
pydantic>=2.0.0
aiohttp>=3.9.0
tenacity>=8.2.0
reportlab
Pillow
arabic-reshaper
python-bidi
nakdimon
requests
